refactor: infra/hub 패키지를 infra/client로 재구성
