+++ src/main/java/com/followMe/delivery_server/delivery/infra/client/HubClient.java
@@ -1,16 +1,13 @@
-package com.followMe.delivery_server.delivery.infra.hub;
+package com.followMe.delivery_server.delivery.infra.client;
 
-import com.followMe.delivery_server.delivery.infra.hub.dto.HubRouteResponse;
+import com.followMe.delivery_server.delivery.infra.client.dto.HubRouteResponse;
 import java.util.UUID;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@FeignClient(
-    name = "hub-service",
-    fallbackFactory = HubClientFallbackFactory.class,
-    primary = false)
+@FeignClient(name = "hub-server", fallbackFactory = HubClientFallbackFactory.class, primary = false)
 public interface HubClient {
-  @GetMapping("/api/hubs/route")
+  @GetMapping("/internal/hubs/route")
   HubRouteResponse getNodes(@RequestParam UUID sourceHubId, @RequestParam UUID vendorId);
 }
