+++ src/main/java/com/followMe/delivery_server/delivery/infra/client/HubClientLocalStub.java
@@ -1,9 +1,9 @@
-package com.followMe.delivery_server.delivery.infra.hub;
+package com.followMe.delivery_server.delivery.infra.client;
 
 import com.followMe.delivery_server.delivery.domain.enums.NodeType;
 import com.followMe.delivery_server.delivery.domain.exception.HubClientUnavailableException;
-import com.followMe.delivery_server.delivery.infra.hub.dto.HubNodeInfo;
-import com.followMe.delivery_server.delivery.infra.hub.dto.HubRouteResponse;
+import com.followMe.delivery_server.delivery.infra.client.dto.HubNodeInfo;
+import com.followMe.delivery_server.delivery.infra.client.dto.HubRouteResponse;
 import java.util.List;
 import java.util.UUID;
 import lombok.extern.slf4j.Slf4j;
@@ -14,22 +14,21 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Primary
 @Component
-@Profile("local")
+@Profile({"local", "local-intellij"})
 public class HubClientLocalStub implements HubClient {
 
   private static final UUID MIDDLE_HUB_ID = UUID.fromString("dddddddd-0000-0000-0000-000000000002");
   private static final String UNKNOWN_SUFFIX = "0000-000000000099";
 
   @Override
-  public HubRouteResponse getNodes(UUID sourceHubId, UUID vendorId) {
-    log.info("[LocalStub] HubClient.getNodes hubId={} vendorId={}", sourceHubId, vendorId);
-    if (sourceHubId.toString().endsWith(UNKNOWN_SUFFIX)
-        || vendorId.toString().endsWith(UNKNOWN_SUFFIX)) {
+  public HubRouteResponse getNodes(UUID hubId, UUID vendorId) {
+    log.debug("[LocalStub] HubClient.getNodes sourceHubId={} vendorId={}", hubId, vendorId);
+    if (hubId.toString().endsWith(UNKNOWN_SUFFIX) || vendorId.toString().endsWith(UNKNOWN_SUFFIX)) {
       throw new HubClientUnavailableException();
     }
     return new HubRouteResponse(
         List.of(
-            new HubNodeInfo(sourceHubId, NodeType.HUB, "출발 허브", 1),
+            new HubNodeInfo(hubId, NodeType.HUB, "출발 허브", 1),
             new HubNodeInfo(MIDDLE_HUB_ID, NodeType.HUB, "중간 허브", 2),
             new HubNodeInfo(vendorId, NodeType.VENDOR, "도착 업체", 3)));
   }
