+++ src/main/java/com/followMe/delivery_server/delivery/infra/client/dto/HubRouteResponse.java
@@ -1,4 +1,4 @@
-package com.followMe.delivery_server.delivery.infra.hub.dto;
+package com.followMe.delivery_server.delivery.infra.client.dto;
 
 import com.followMe.delivery_server.delivery.domain.Node;
 import com.followMe.delivery_server.delivery.domain.exception.InvalidNodeInformationException;
